requirements.txt
google-genai
google-ai-generativelanguage==0.6.15
langchain-google-genai
langchain-community
langchain-tavily
langchain
langgraph
python-dotenv
ipython
tavily
tavily-python
ddgs
bs4
selenium
webdriver-manager
aiosqlite
langgraph-checkpoint-sqlite
gradio